requirements.txt
bcrypt==3.1.7

email-validator==1.1.1

Flask==1.1.2

Flask-Bcrypt==0.7.1

Flask-Login==0.5.0

Flask-Mail==0.9.1

Flask-SQLAlchemy==2.4.2

Flask-WTF==0.14.3

Pillow==7.1.2

pycparser==2.20

python-dotenv==0.13.0

SQLAlchemy==1.3.17